refactor: Criando molecula para o solicitations menu
